split out each field and added unit tests
